+++ app/src/main/res/drawable/ic_find.xml
@@ -1,0 +1,28 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="330dp"
+    android:height="50dp"
+    android:viewportWidth="330"
+    android:viewportHeight="50">
+  <path
+      android:strokeWidth="1"
+      android:pathData="M9,1L321,1A8,8 0,0 1,329 9L329,41A8,8 0,0 1,321 49L9,49A8,8 0,0 1,1 41L1,9A8,8 0,0 1,9 1z"
+      android:fillColor="#ffffff"
+      android:strokeColor="#E4E4EB"/>
+  <path
+      android:pathData="M304,32C308.418,32 312,28.418 312,24C312,19.582 308.418,16 304,16C299.582,16 296,19.582 296,24C296,28.418 299.582,32 304,32Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M314,34L309.65,29.65"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M28.36,30.936H21.784V29.08H28.36V30.936ZM28.36,28.056H21.784V26.248H20.488V31.992H29.672V26.248H28.36V28.056ZM21.768,24.056C20.344,24.056 19.32,23.128 19.32,21.816C19.32,20.504 20.344,19.592 21.768,19.592C23.192,19.592 24.216,20.504 24.216,21.816C24.216,23.128 23.192,24.056 21.768,24.056ZM28.344,17.848V21.256H25.416C25.144,19.592 23.704,18.504 21.768,18.504C19.624,18.504 18.072,19.848 18.072,21.816C18.072,23.8 19.624,25.144 21.768,25.144C23.736,25.144 25.176,24.024 25.432,22.328H28.344V25.544H29.672V17.848H28.344ZM36.743,21.56H39.399V20.472H36.759V18.344H35.495V20.472H32.839V21.56H35.495V22.248C35.495,24.6 34.391,27.112 32.455,28.264L33.223,29.24C34.631,28.392 35.639,26.792 36.135,24.968C36.663,26.696 37.687,28.152 39.079,28.936L39.831,27.944C37.895,26.904 36.743,24.536 36.743,22.248V21.56ZM40.615,23.496H38.423V24.584H40.615V31.432H41.863V18.168H40.615V23.496ZM43.495,17.832V32.168H44.759V17.832H43.495ZM52.681,18.744V19.816H57.769C57.497,22.248 55.369,24.12 51.961,25.096L52.505,26.12C56.617,24.952 59.161,22.376 59.161,18.744H52.681ZM55.641,27.672H62.329V30.92H55.641V27.672ZM54.345,31.976H63.609V26.632H54.345V31.976ZM62.297,17.832V21.528H59.177V22.6H62.297V26.024H63.609V17.832H62.297ZM73.704,24.632C71.848,23.896 70.792,22.232 70.792,20.616V18.68H69.528V20.616C69.528,22.248 68.408,24.136 66.536,24.984L67.24,25.992C68.632,25.368 69.656,24.152 70.168,22.792C70.68,24.024 71.672,25.08 73,25.624L73.704,24.632ZM68.984,28.36H77.32V32.184H78.632V27.304H68.984V28.36ZM77.368,17.848V21.576H75.544V18.088H74.296V26.424H75.544V22.648H77.368V26.52H78.632V17.848H77.368Z"
+      android:fillColor="#7D89A0"/>
+</vector>
