+++ app/src/main/res/drawable/ic_home.xml
@@ -1,0 +1,19 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="40dp"
+    android:viewportWidth="24"
+    android:viewportHeight="40">
+  <path
+      android:pathData="M0,0h24v24h-24z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M3,9L12,2L21,9V20C21,20.5304 20.7893,21.0391 20.4142,21.4142C20.0391,21.7893 19.5304,22 19,22H5C4.4696,22 3.9609,21.7893 3.5858,21.4142C3.2107,21.0391 3,20.5304 3,20V9Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M16.621,27.263H12.5V26.145H11.434V27.263H7.235V28.095H16.621V27.263ZM9.029,35.297H14.905V37H9.029V35.297ZM7.989,37.806H15.945V34.491H7.989V37.806ZM9.315,30.201C9.315,29.629 10.264,29.304 11.967,29.304C13.644,29.304 14.593,29.629 14.593,30.201C14.593,30.786 13.644,31.111 11.967,31.111C10.264,31.111 9.315,30.786 9.315,30.201ZM12.5,32.736V31.852C14.554,31.774 15.724,31.202 15.724,30.201C15.724,29.122 14.359,28.55 11.967,28.55C9.549,28.55 8.197,29.122 8.197,30.201C8.197,31.202 9.354,31.774 11.434,31.852V32.736H6.689V33.581H17.219V32.736H12.5Z"
+      android:fillColor="#000000"/>
+</vector>
