+++ app/src/main/res/drawable/ic_map_icon.xml
@@ -1,0 +1,25 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M3,17a9,4 0,1 0,18 0a9,4 0,1 0,-18 0z"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:pathData="M18,8.7273C18,13.1818 12,17 12,17C12,17 6,13.1818 6,8.7273C6,7.2083 6.6321,5.7516 7.7574,4.6775C8.8826,3.6034 10.4087,3 12,3C13.5913,3 15.1174,3.6034 16.2426,4.6775C17.3679,5.7516 18,7.2083 18,8.7273Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#ffffff"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M12,11C13.1046,11 14,10.1046 14,9C14,7.8954 13.1046,7 12,7C10.8954,7 10,7.8954 10,9C10,10.1046 10.8954,11 12,11Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
