+++ app/src/main/res/drawable/ic_my.xml
@@ -1,0 +1,26 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="40dp"
+    android:viewportWidth="24"
+    android:viewportHeight="40">
+  <path
+      android:pathData="M12.1943,37V27.8403H10.8677L7.7319,35.4829H7.6304L4.4946,27.8403H3.168V37H4.2344V30.0049H4.3106L7.1987,37H8.1636L11.0518,30.0049H11.1279V37H12.1943ZM18.1992,37V33.0581L21.4492,27.8403H20.1479L17.6787,31.9092H17.5771L15.1079,27.8403H13.8066L17.0566,33.0581V37H18.1992Z"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M0,0h24v24h-24z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M20,21V18.6667C20,17.429 19.5786,16.242 18.8284,15.3668C18.0783,14.4917 17.0609,14 16,14H8C6.9391,14 5.9217,14.4917 5.1716,15.3668C4.4214,16.242 4,17.429 4,18.6667V21"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M12,11C14.2091,11 16,9.2091 16,7C16,4.7909 14.2091,3 12,3C9.7909,3 8,4.7909 8,7C8,9.2091 9.7909,11 12,11Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+</vector>
