+++ app/src/main/res/drawable/item_color.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="#132DB3" android:state_checked="true" />
+    <item android:color="#FFFFFF" />
+</selector>
